Refactor(SG) : entity 폴더 이동 #7
