+++ src/main/java/com/example/sinsungo/chat/entity/ChatRoom.java
@@ -1,5 +1,5 @@
 package com.example.sinsungo.chat.entity;
-import com.example.sinsungo.entity.TimeStamped;
+import com.example.sinsungo.common.entity.TimeStamped;
 import com.example.sinsungo.user.User;
 import jakarta.persistence.*;
 import lombok.Getter;
