+++ src/main/java/com/example/sinsungo/ingredient/entity/Ingredient.java
@@ -1,6 +1,6 @@
 package com.example.sinsungo.ingredient.entity;
 
-import com.example.sinsungo.entity.TimeStamped;
+import com.example.sinsungo.common.entity.TimeStamped;
 import jakarta.persistence.*;
 import lombok.Getter;
 import org.hibernate.annotations.DynamicUpdate;
