+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:mysql://localhost:3306/sinsungo
 spring.datasource.username=root
-spring.datasource.password=seung6774!
+spring.datasource.password=bkj09270!
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.jpa.hibernate.ddl-auto=update
